refactor(queue_drag): rightbar queue drag handling
